Implemented DIFFABLE data sources instead of regular ones for both Categories and Warranties collection views in HomeWarrantiesListVC.
Started working again on Categories. Improved UI, but still not functional. 

Will first and foremost try to move all things relative to Categories in a new, dedicated file "CategoriesVC". Aim is to make HomeWarrantiesListVC shorter. In that sense, an extension of UIViewController containing "start()" and "remove()" was added to be able to add the child CategoriesVC to the view.

+++ WhereIsMyWarranty/Extensions/CornerRadius.swift
@@ -9,14 +9,14 @@ import UIKit
 
 // ⬇︎ Rounds the corresponding item's corners, based on a given radius
 extension UICollectionViewCell {
-    func roundingCellCorners(radius: Int) {
+    func roundingCellCorners(radius: Float) {
         self.layer.masksToBounds = true
         self.contentView.layer.cornerRadius = CGFloat(radius)
     }
 }
 
 extension UIView {
-    func roundingViewCorners(radius: Int) {
+    func roundingViewCorners(radius: Float) {
         self.layer.masksToBounds = true
         self.layer.cornerRadius = CGFloat(radius)
     }
